Fixed (really this time): Node task fails if Node is installed globally (i.e. node.exe is in the PATH).
When Node is installed globally instead of being managed by NVM, it sets up the environment such that any Node command or script that references node.exe without using an absolute path will get the first version of Node in the user's path. So, prepend the path to the version of Node we want to use to the user's path before running any node commands.