Add support for including third-party packages/tools in a package.
In Node.js applications, we should always package and deploy the node_modules directory as-is. Currently, this requires the application's whitelist to include all the files from the node_modules directory. Update the package task to support including these kind of third-party items so they don't need to be included on the whitelist.