PublishPowerShellModule task improvements.
Turns out, we don't need to trust any repositories or install any package providers. Publish-Module needs nuget.exe in the PATH. If it can't find nuget.exe, it attempts to install it, which fails when running non-interactively (i.e. under a build server). To fix, we temporarily add WhsCI's nuget.exe to the PATH.