Redirect all verbose messages to null when importing a module
The "Parallel" task displays all verbose messages that were written
during the background job, regardless of the "$VerbosePreference" value.
Therefore when importing PowerShell modules we need to redirect all
verbose messages to null to prevent the import messages from entering
the verbose stream and being returned at the end of the Parallel task.