Migrating to single-It-per-Describe test pattern.
Pester tests no longer use external test fixtures. Instead, each test
writes its own fixture.

Eventually, I hope that Invoke-Pester supports running individual `It`
blocks, so we can switch to the community pattern of multiple `It`
blocks per `Describe`. It turns out that Pester's TestDrive is setup and
torn down for each `Describe` block, so almost all our tests were
written with the assumption that they got a fresh test drive
uncontaminated by other tests. If and when we move to a
multiple-It-per-Describe pattern, tests need to setup their own area in
the test drive. So, I created a `New-WhiskeyTestRoot` function to do
that. Every test fixture I touched I also migrated to use this new
function to make migration to a future version of Pester that supports
running single `It` blocks.