Whiskey now auto-imports PowerShell modules for tasks.
The reason for this commit is because tasks were failing because the
PackageManagement and PowerShellGet modules weren't getting
auto-installed on the build server. There was a bug where we weren't
bootstrapping them in the right place.

Let's make callers of Whiskey's various *-WhiskeyPowerShellModule be
explicit about where the build root is. No more guessing or hoping it's
always the current directory.

When we can't find a PowerShell module, don't assume we're using the
PSGallery. Instead, show the user what repository's are configured.