Create a module of tasks to use in tests that don't test specific tasks.
I changed the signature of the PowerShell task and that caused a bunch
of the Invoke-WhiskeyTask tests to fail, as they were dependent on that
task. A bunch of the tests were also creating their own tasks in the
global scope and then deleting them. A module seems like a better place
for these tasks so they can be imported/removed easily.