Pester4 task improvements:
* Use named parameters instead of the TaskParameter hashtable.
* Give users the option to *not* run tests in a background job. Some
testing tools don't like that.
* Allow users to pass custom arguments to Pester via an `Arguments`
property.

I tried to get Pester4 to run tests not in a background job by default,
but there was too much weirdness. Plus, it makes sense that tests should
run in an isolated process. In trying to get the Pester4 task to run
tests in scope, I made some changes to test fixtures that are
beneficial, so I'm keeping them.

Created a new `Write-WhiskeyObject` function so tasks can more-easily
write objects to the output in a recognizable format.