Improving how Whiskey resolves paths.
Moving all tasks that require resolved paths to use a `ValidatePath`
attribute instead (except `DotNet` for backwards-compatability). It's
good to have all tasks behave the same way.

Resolved paths are now returned as relative paths instead of absolute
paths. Relative paths look better and take up less space in Whiskey's
output. Also, this reduces the risk that someone will hit the Windows
max command line size.

Renamed Resolve-WhiskeyTaskPathParameter to be Resolve-WhiskeyTaskPath
so that it now becomes the default path resolver for tasks (and other
functions) that don't want to use the `ValidatePath` attribute. The
old function is still in use by some deprecated tasks, so it sticks
around (sans tests).

Paths should be resolved relative to a task's working directory, not
always the build root. We think this is the behavior most users would
expect. It also helps to make tasks in whiskey.yml more terse.