Users should be able to publish a PowerShell module with their deployment tool of choice instead of using Whiskey.
The `PublishPowerShellModule` task now, by default, publishes a .nupkg package of a PowerShell module to a build's
output directory. This .nupkg can then be published with another tool to a NuGet-based PowerShell feed.

Update Whiskey's own build pipeline to no longer deploy during the build. We'll now use AppVeyor deployments to deploy
on-demand.