Fixed: PublishPowerShellScript tests failing because an old version of PowerShellGet is loaded.
The functions used in test to add/remove paths from PSModulePath had bugs in them. The path to Whiskey's PSModules
directory, which is first in the list because it contains the correct version of PowerShellGet, was getting removed by
PublishBitbucketServerTag and PublishBuildMasterPackage tests. Tests now use Whiskey's internal
Register-WhiskeyPSModulePath and Unregister-WhiskeyPSModulePath functions.