Enhancements to the InstallNode task.
To resolve the version to install, the task now uses, in this order:

* a Version property on the task in the whiskey.yml.
* a .node-version file, if it exists.
* a 'whiskey.node' property in package.json file.

It will also pin to a specific version of NPM by looking for the version, in this order:

* an NpmVersion property on the task in whiskey.yml file.
* a `whiskey.npm` property in package.json file.

Add the path to where Node.js get installed to the PATH environment variable so builds can install Node.js and then run commands against.

Allow customizing the CPU architecture of the Node.js package to download.