Allow tasks to have properties that currently match reserved, common properties.
We forsee a future where task authors might want task properties that have the same name as Whiskey's current builtin,
common properties. Common property names are reserved at the moment, unavailable to task authors. The fix is to prefix
Whiskey's common properties and then not allow tasks to have properies that begin with that prefix.