Adding support for snake_case task and property names.
We want to move to using snake_case naming conventions by default. Step 1 is to allow users to update their whiskey.yml
files to use it first, before updating Whiskey's internals to require it. As part of this, renamed Whiskey's common
properties to be snake cased by default. This change should mostly be silent to users.