Fixed: InstallNodeJs tests take a really, really, really long time on Linux.
After each Context block, Pester deletes the contents of the test drive, but does it one item at a time. The tests hang for 30 seconds locally and minutes on the build server while each item is deleted. We cache across tests so don't want anything deleted after a Context block runs. So we have to use our own temp directory.